@alrund 2022-02-26
benchstat -delta-test none old new
name              old time/op    new time/op    delta
GetDomainStat-16     349ms ± 0%       6ms ± 0%   -98.14%

name              old alloc/op   new alloc/op   delta
GetDomainStat-16     133MB ± 0%       0MB ± 0%  -100.00%

name              old allocs/op  new allocs/op  delta
GetDomainStat-16     1.70M ± 0%     0.00M ± 0%  -100.00%